Undid previous patch, the buffs were too stronk. sorry about that, back to competitive play! (buy our skins)

+++ src/main.cpp
@@ -42,7 +42,7 @@ IRsendNECx sender;
 LiquidCrystal_I2C lcd(0x27,20,4);  // set the LCD address to 0x27 for a 16 chars and 2 line display
 
 
-int score = 0;
+long score = 0;
 
 const int maxHealth = 5;
 int currentHealth = maxHealth;
